Do not track configuration inputs when inside the ValueSource
The ValueSource's value is a part of the configuration cache fingerprint
and it is recomputed each time the build runs, even when it runs from
cache. There is no need to additionally track inputs as only the value
of the ValueSource matters in the end.

For example, this allows users to implement complex filtering of the
available system properties without making every property an input to
the configuration phase.